Make path manipulation play nice with symlinks by deferring symlink resolution
pathlib.Path.resolve() would follow and resolve symlinks, which is
usually not what we want. In most cases we want to defer symlink
resolution.
This change makes all places use pathlib.Path.absolute(). It does not
normalize the path like resolve, but in all cases we don't care if it
is normalized. It is also probably good in general that downstream
functionality is not brittle by accepting only normalized paths.
When doing path manipulation we may care about the surrounding
directory structure, which may be a collection of multiple symlinks,
so doing a symlink resolution would put you in a different directory
structure, which is probably not the intent.