Add Flux transformer benchmarking
Add also some more general functionality to check benchmark results
against baseline results. This requires the Google Benchmark compare.py
tool that is not a part of the pip package. That is why I added the
repo as a git submodule. This tool does a statistical comparison
between benchmarks with proper p-value calculation. I don't think we
should roll out our own.

Adds a new nightly CI job that should contain nightly tests and
benchmarks that are not in their own category like Llama is now.