Add deepseek v3 model support
This adds an e2e implementation of deepseek for prefill. Additional
features were required by the MoE block, specifically surrounding how
gate scores are computed and including mixture of experts.

A latent attention implementation is also included though this is a
reference implementation that does not take advantage of the latent
space during decode.