[sharktank][kernels] restores unmasked attn functionality
previous patch erroneously rerouted the flash attention
function inside of attention_impls.py directly to the custom kernel,
skipping the important step of managing the scales properly.
This restores the previous functionality. Something similar probably
needs to be done for the masked kernel, and thats why the previous
patch renamed flash_attention() to prepare_args(), with the intention
of factoring out the code needed for both.