[sharktank] make tests optionally more verbose to log start/finish
Occasionally have hanging tests in CI and it would be nice to know
which test is the culprit.

Here all tests are made to inherit from
`sharktank.utils.testing.TestCase`, which has a test fixture to log the
start and end of all each test.
This will enable us to see which test started and did not finish even
when running multiple tests in parallel.