[sharktank] Add reduce_scatter and split ops
Adds reduce_scatter for an UnreducedTensor, which would be required for
tensor-sharding the latent attention preprocessing.

A Split op is also added as it is used in the implementation of
reduce_scatter.

The sharded_sum op is augmented with a new root_rank argument.
It specifies the rank of the destination device within the sharded
tensor. It has the same meaning as in MPI.