[sharktank] xfail with support for regex matching in the error message (#1521)
This wraps the pytest.mark.xfail decorator into a new decorator.
pytest.mark.xfail does not support matching on the error message, but
sometimes we need to be more precise on why we expect a failure. One
example is when specifying what compiler error is expected. Just the
exception type is not enough.

```
@xfail(raises=MyError, strict=True, match="my message")
@test_something():
    raise MyError("my message")
```