Add --use-attention-mask flag to eager mode (#1542)
IREE version of models could be run with or without the
`--use-attention-mask` flag, with the default being without the flag.
The eager mode execution could only be run **with** the flag, and there
was no way to disable it.

This PR:
- Adds the flag to eager runs
- Sets the default to enabled so that current behaviour is maintained

This is useful for testing differences model results, since the
resulting mask is slightly different with or without the flag, as have
been observed in the past.

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>