[sharktank] Allow for non-sharded tensor args in device verification wrapper and exclude some ops
There was a torch export dynamic shapes constraint violation problem
with ops.trace_tensor. It was called on a unsharded tensor, which
caused use to loop over dimension 0 of the tensor in
sharded_wrap_override.transfer_n_pin.func_wrapper. This caused torch
tracing to introduce unwanted constraints on the dynamic dims.

In any case we don't want `unshard` and `trace_tensor` to be wrapped by
the verifier. `unshard` as a single tensor arg op and it is a
transferring function. `trace_tensor` is a an op with tensor args that
can be sharded differently.

There is the question of whether we should allow mixing of unsharded
and sharded tensors args in an op. But that should be a concern of the
underlying op, not the wrapper.