[sharktank] fix conversion between torch.Tensor and IREE device array
Recently we introduced a regression #1293 in the conversion where we
used torch.Tensor.to instead of torch.Tensor.view to do dtype
reinterpretation.

PR #1172 introduced handling of f8, but left taking a view of f8 as int16,
which is incorrect.

Here the approach is generalized using various maps to reinterpret as
the appropriate dtypes.

Here are introduced tests for roundtrip conversion of various dtypes.