[sharktank] refactor pipeline-parallel FFN example
The main modeling structure is maintained where layers/blocks are mapped
to pipeline stages and pipeline stages are mapped to device sets.
This looks like the most natural way to represent this.

Utilize method _prepare_pipeline_parallel_block_args to prepare
the arguments of a layer/block.

Write actually the pipeline parallelized theta and not the
original before pipelining. This required adding device
writing/reading when saving/loading sharded tensors.

Refactor pipeline-parallelisation of the theta to not use the function for LLMs
directly to avoid having to include unused parameters like token_embd, etc.

Make 2 tests. One where pipeline-parallelism is combined with tensor-parallelism
and one where it is without tensor-parallelism.