Bump iree to `3.7.0rc20250822` and kernels for matmul_transpose change (#2083)
Interface changes in
[IREE-21619](https://github.com/iree-org/iree/pull/21619), which require
a small update in the shortfin code.

In that PR, @benvanik states that "Behavior is defaulted to what it is
today before these additions...", so I followed what seemed to be the
default flags throughout that PR to maintain current behavior (i.e.
[here](https://github.com/iree-org/iree/pull/21619/files#diff-289c6144e8b6d353bc4e1dba6a206d8b2da145ffa66f5e659cf1b6c2345103cfR386)).

Could use a double-check to make sure that this is in fact maintaining
current behavior on the `shortfin` side.

Also includes @vivekkhandelwal1's changes to fix the `matmul_transpose`
op, merged from #2055

---------

Signed-off-by: Vivek Khandelwal <vivekkhandelwal1424@gmail.com>
Co-authored-by: Vivek Khandelwal <vivekkhandelwal1424@gmail.com>