[sharktank] Make torch SDPA the default attention impl
Currently "sharktank" is the default attention implementation, but this has the
problem that it unconditionally always returns a f32 tensor.
This diverges from the semantics of the torch operation, where the return dtype is
dependent on the operands' dtypes.
It is strongly desirable for the default to do as torch does as it has a well
documented behaviour. In general we should try hard for our defaults to not diverge
from torch's semantics as this will cause confusion when porting models in the future.
This actually already caused a problem with the Flux transformer, where downstream ops
after attention would be in f32 instead of the desired bf32.

The change here relies on the resolution order of overrides, if no impl is specified then
"torch" will be picked before "sharktank".