Replace `use_hf` with `interleave_emb` for rotary embedding
Use_hf is ambiguous and should be identified in the IRPA file. Changed
to set a `rotary.interleave_emb` in the hyperparameters. Quark and
huggingface import appends this information to the configuration as the
source will tell the layout of the rotary embedding.

Include a tool for updating properties of existing irpa files.