[FusilliPlugin] Stub plugin (`fusilli-plugin` 2 of N)
This PR adds "all the boring bits" of the fusilli plugin. It adds a
basic engine plugin that hipDNN can load, which ensures that all of the
required API definitions are present, and those APIs called as part of
load function. The PR also tests for plugin loading, logging, and basic
functionality.