[FusilliPlugin] Graph execution (fusilli-plugin 4 of 4)
This PR implements `hipdnnEnginePluginExecuteOpGraph` which enables plan
execution. We can now complete the initial plugin goal of e2e run of
single node ConvFprop graph.

Note: a simple e2e integration test is added. This should be expanded to
a parameterized matrix of tests, once we're on the latest hipDNN
[example](https://github.com/ROCm/rocm-libraries/blob/develop/projects/hipdnn/plugins/miopen_legacy_plugin/integration_tests/IntegrationGpuConvForward.cpp). Tickets: [#2363](https://github.com/nod-ai/shark-ai/issues/2363), [#2375](https://github.com/nod-ai/shark-ai/issues/2375).