Address PR comment and fix tests
Make tests actually use the correct KV cache dtype.

Fix the name of the nightly job and a step to run integration tests.

Put the test for Llama 3.1 405B in the integration tests directory.

Make LlmPerplexityCompare work directly with LlmInstance instead of
TorchInstance/IreeInstance and make it expose rtol and atol.

Make index_copy_ also do a f8 workaround when on eager GPU.

Move PlanarQuantizedTensor.to impl into separate PR.

Make the tests run on CPU due to inability of custom kernels to run with
current wheel of PyTorch ROCm.