[sharktank] Change scheduler interface to allow for more broad uses of chunked prefill
Make the selection function return `list[Any]` for a batch. This for example allows us
to propagate the logits instead of doing a selection.

Make the run Scheduler.run function return a list of selections for each request id.
This is needed to return the chunk results. It is not appropriate to do collection and
stitching of results in the scheduler as its puropse is to run the tasks. It is not
concerned with breaking up a request into tasks, so it should not be concerned of doing
any sort of reduction.