[Fusilli] Use per github action run temp dir for cache.
I've seen race conditions on the compile cache on the fusilli samples
when running in CI. The only place that the race seems to be possible is
between multiple CI runs on the same machine both touching `/tmp`.
There's two `SECTIONS` in each fusilli sample (for CPU and AMDGPU
back-end) but the docs that I've found indicate that `SECTIONS` run
serially within one test executable not in parallel, so even though the
graph name is shared between the two sections, there shouldn't be a
race. This PR sets `FUSILLI_CACHE_DIR` to a per-github-action-job temp
directory.