[DO NOT REVIEW] CI testing for `ILLEGAL` error occasionally seen in CI
When diagnosing this issue, some sources point to the possibility of an
architecture mismatch, i.e. we generate code that assumes a particular
instruction is available, and if said instruction isn't available we get
a `SIGILL`and the `ILLEGAL` error from CTest we've seen. Without `--iree-llvmcpu-target-cpu=host` IREE will not attempt to use
any advanced CPU arch features making an arch mismatch impossible. If
we see the same `ILLEGAL` error without the `--iree-llvmcpu-target-cpu=host` flag, we can be
reasonably assured that arch mismatch isn't the culprit.